requirements.txt
Flask
torch
torchvision
torch_geometric
pandas
numpy
scikit-learn
Pillow
tqdm
openpyxl